#327 Refactor: export constants to index, and change path import
